arm64: fix selcall call data for return of aggregate in memory
The no-op `copy R0` is necessary in order to trigger dopm in spill.c
and rega.c, which assume that a call is always followed by one or
more copies from registers. However, the arm64 ABI does not actually
return the caller-passed pointer as in x86_64. This causes an
assertion failure

	qbe: aarch64: Assertion failed: r == T.rglob || b == fn->start (spill.c: spill: 470)

for the following test program

	type :t = { l 3 }
	function $f() {
	@start.1
	@start.2
		%ret =:t call $g()
		ret
	}

The assertion failure only triggers when the block containing the
call is not the first block, because the check is skipped for the
first block (since some registers may have been used for arguments).

To fix this, set R0 in the call data so that spill/rega can see
that this dummy "return" register was generated by the call. This
matches qbe's existing behavior when the function returns void,
another case where no register is used for the function result.