amd64: optimize loading +0 into floating point registers
Loading +0 into a floating point register can be done using pxor or
xorps instructions. Per [1], we went with pxor because it can run on all
vector ALU ports, even if it's one byte longer.

To implement it, we special case fixarg to emit Ocopy when the value is
+0, and emitins to treat Ocopy for floating point args specially. Since
0. == -0., we can't check if bits.d or bits.f are equal to 0. To avoid
requiring signbit(), we inspect bits.i directly; this assumes the bits
union is always zero-initialized.

[1] https://stackoverflow.com/questions/39811577/does-using-mix-of-pxor-and-xorps-affect-performance/39828976