rv64: sign-extend 32-bit values in 64-bit registers
The riscv64 ABI requires that 32-bit values be passed sign-extended
(regardless of the sign) in 64-bit registers. This convention allows
64-bit signed/unsigned compare to be used on 32-bit values, so no
32-bit compare instructions exist.

However, due to QBE's subtyping feature, we may end up with l
operands in w positions. This is problematic for instructions that
do not discard the upper 32 bits of their operands (such as slt or
sltu), or when passing 32-bit arguments to functions.

To ensure that we maintain this invariant, detect subtyping in
fixarg() and insert sext.w instructions as needed. Also, when loading
a 32-bit value into a w register, use lw regardless of whether the
original instruction was loaduw or loadsw.

As previously mentioned, some instructions (such as the 32-bit
arithmetic instructions) don't require the sign extension since
they discard the upper bits, but this is saved for a future
optimization.